+++ CHANGELOG.md
@@ -10,3 +10,9 @@
 ### Fixed
 #### Closing issue (hardening)
 - Add missing async result handler call on context closing hook.
+
+## [0.0.3]
+### Fixed
+#### Handle client is busy (hardening)
+- Cover the case if the user is trying to execute too many commands at once -> ConnectionPoolTooBusyException. 
+A RedisHeimdallException (with CLIENT_BUSY reason) exception will get thrown, and the reconnecting process will NOT get started.